created specific folder to deal with Parameter objects related functions and type def
