+++ pymodaq/daq_utils/parameter/__init__.py
@@ -1,0 +1,1 @@
+from pyqtgraph.parametertree import parameterTypes
